requirements.txt
pyyaml
ttkbootstrap